requirements.txt
python-aiml
pandas
nltk
scikit-learn
numpy
Pillow
opencv-python 